+++ src/main/java/seedu/address/logic/parser/NextOfKinCommandParser.java
@@ -6,13 +6,19 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_REMARK;
-import static seedu.address.model.person.Phone.isValidPhone;
+import static seedu.address.model.person.NextOfKin.MESSAGE_EMAIL_CONSTRAINTS;
+import static seedu.address.model.person.NextOfKin.MESSAGE_PHONE_CONSTRAINTS;
+import static seedu.address.model.person.NextOfKin.MESSAGE_REMARK_CONSTRAINTS;
+import static seedu.address.model.person.NextOfKin.isValidEmail;
+import static seedu.address.model.person.NextOfKin.isValidPhone;
+import static seedu.address.model.person.NextOfKin.isValidRemark;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.EditPersonDescriptor;
 import seedu.address.logic.commands.NextOfKinCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+
 
 /**
  * Parses input arguments and creates a new EditCommand object
@@ -39,14 +45,15 @@ public class NextOfKinCommandParser implements Parser<NextOfKinCommand> {
         }
 
         EditPersonDescriptor editPersonDescriptor = new EditPersonDescriptor();
-        if (!(argMultimap.arePrefixesPresent(PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_REMARK))) {
+        if (!(argMultimap.arePrefixesPresent(PREFIX_NAME, PREFIX_PHONE, PREFIX_REMARK))) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, NextOfKinCommand.MESSAGE_USAGE));
         }
         try {
             ParserUtil.parseNextOfKin(argMultimap.getValue(PREFIX_NAME),
-                    argMultimap.getValue(PREFIX_PHONE),
-                    argMultimap.getValue(PREFIX_EMAIL),
-                    argMultimap.getValue(PREFIX_REMARK)).ifPresent(editPersonDescriptor::setNextOfKin);
+                        argMultimap.getValue(PREFIX_PHONE),
+                        argMultimap.getValue(PREFIX_EMAIL),
+                        argMultimap.getValue(PREFIX_REMARK)).ifPresent(editPersonDescriptor::setNextOfKin);
+
         } catch (IllegalValueException ive) {
             throw new ParseException(ive.getMessage(), ive);
         }
@@ -56,11 +63,22 @@ public class NextOfKinCommandParser implements Parser<NextOfKinCommand> {
             throw new ParseException(NextOfKinCommand.MESSAGE_NOT_EDITED);
         }
 
-        if (isValidPhone(argMultimap.getValue(PREFIX_PHONE).get())){
-
+        if (!isValidPhone(argMultimap.getValue(PREFIX_PHONE).get())) {
+            throw new ParseException(MESSAGE_PHONE_CONSTRAINTS);
         }
+
+        if (argMultimap.getValue(PREFIX_EMAIL).isPresent()) {
+            if (!isValidEmail(argMultimap.getValue(PREFIX_EMAIL).get())) {
+                throw new ParseException(MESSAGE_EMAIL_CONSTRAINTS);
+            }
+        }
+
+        if (!isValidRemark(argMultimap.getValue(PREFIX_REMARK).get())) {
+            throw new ParseException(MESSAGE_REMARK_CONSTRAINTS);
+        }
+
 
         return new NextOfKinCommand(index, editPersonDescriptor);
     }
-//@@author
+    //@@author
 }
